// doggywood/src/app/models/customer.ts
export class Customer {
    id: number;
    firstName: string;
    lastName: string;
    phone: string;
    email: string;
    password: string;
    cusUrl: string;

    // constructor(
    //     id: number,
    //     firstName: string,
    //     lastName: string,
    //     phone: string,
    //     email: string,
    //     password: string,
    //     cusUrl: string
    // ) {
    //     this.id = id;
    //     this.firstName = firstName;
    //     this.lastName = lastName;
    //     this.phone = phone;
    //     this.email = email;
    //     this.password = password;
    //     this.cusUrl = cusUrl;
    // }

}